Validate ndjson file

Given a file of streaming JSON (ndjson) this function reads in the records and validates that they are all legal JSON records. If the verbose parameter is TRUE and errors are found, the line numbers of the errant records will be displayed.

validate(path, verbose = FALSE)

Arguments

path

path to file (supports "gz" files)

verbose

display verbose information (filename and line numbers with bad records)

Value

logical

References

http://ndjson.org/

Examples

f <- system.file("extdata", "test.json", package="ndjson")
validate(f)

#> [1] TRUE
#> 

gzf <- system.file("extdata", "testgz.json.gz", package="ndjson")
validate(gzf)

#> [1] TRUE
#>